requirements.txt
streamlit
langchain
openai==0.28.0
faiss-cpu
python-dotenv